LPS-28031 Refactor for plugins
