+++ portal-impl/src/com/liferay/portlet/dynamicdatalists/util/FreeMarkerTemplateParser.java
@@ -14,12 +14,12 @@
 
 package com.liferay.portlet.dynamicdatalists.util;
 
+import com.liferay.portal.kernel.template.StringTemplateResource;
 import com.liferay.portal.kernel.template.TemplateContextType;
 import com.liferay.portal.kernel.template.TemplateManager;
 import com.liferay.portal.kernel.template.TemplateManagerUtil;
 import com.liferay.portal.kernel.template.TemplateResource;
 import com.liferay.portal.kernel.templateparser.TemplateContext;
-import com.liferay.portal.template.StringTemplateResource;
 import com.liferay.portal.util.PropsValues;
 import com.liferay.util.ContentUtil;
