+++ portal-impl/src/com/liferay/portlet/journal/util/VelocityTemplateParser.java
@@ -15,6 +15,7 @@
 package com.liferay.portlet.journal.util;
 
 import com.liferay.portal.kernel.io.unsync.UnsyncStringWriter;
+import com.liferay.portal.kernel.template.StringTemplateResource;
 import com.liferay.portal.kernel.template.Template;
 import com.liferay.portal.kernel.template.TemplateContextType;
 import com.liferay.portal.kernel.template.TemplateManager;
@@ -27,7 +28,6 @@ import com.liferay.portal.kernel.templateparser.TransformException;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.xml.Element;
-import com.liferay.portal.template.StringTemplateResource;
 import com.liferay.portal.template.TemplateResourceParser;
 import com.liferay.portal.util.PropsValues;
 import com.liferay.util.ContentUtil;
