+++ portal-web/docroot/html/common/init.jsp
@@ -89,6 +89,7 @@ page import="com.liferay.portal.kernel.servlet.SessionErrors" %><%@
 page import="com.liferay.portal.kernel.servlet.SessionMessages" %><%@
 page import="com.liferay.portal.kernel.servlet.StringServletResponse" %><%@
 page import="com.liferay.portal.kernel.staging.LayoutStagingUtil" %><%@
+page import="com.liferay.portal.kernel.template.StringTemplateResource" %><%@
 page import="com.liferay.portal.kernel.upload.UploadException" %><%@
 page import="com.liferay.portal.kernel.util.ArrayUtil" %><%@
 page import="com.liferay.portal.kernel.util.BooleanWrapper" %><%@
@@ -151,7 +152,6 @@ page import="com.liferay.portal.service.permission.LayoutPrototypePermissionUtil
 page import="com.liferay.portal.service.permission.LayoutSetPrototypePermissionUtil" %><%@
 page import="com.liferay.portal.service.permission.PortletPermissionUtil" %><%@
 page import="com.liferay.portal.struts.StrutsUtil" %><%@
-page import="com.liferay.portal.template.StringTemplateResource" %><%@
 page import="com.liferay.portal.theme.PortletDisplay" %><%@
 page import="com.liferay.portal.theme.ThemeDisplay" %><%@
 page import="com.liferay.portal.util.CookieKeys" %><%@
